Enable TensorPipe's SHM transport
The SHM transport uses shared-memory-backed ringbuffers to transfer small payloads between processes on the same machine.

It was disabled in v1.6 due to a CMake mishap but we've since realized that it also doesn't work that well in docker and other setups. Enabling it here to see whether CircleCI fails.

Differential Revision: [D23814828](https://our.internmc.facebook.com/intern/diff/D23814828/)

ghstack-source-id: 119985580
Pull Request resolved: https://github.com/pytorch/pytorch/pull/50760